+++ src/libtrx/include/libtrx/game/game.h
@@ -19,6 +19,8 @@ bool Game_IsBonusFlagSet(GAME_BONUS_FLAG flag);
 
 void Game_SetIsLevelComplete(bool is_complete);
 bool Game_IsLevelComplete(void);
+void Game_FadeToBlack(int32_t duration);
+void Game_DrawFade(void);
 
 extern bool Game_Start(const GF_LEVEL *level, GF_SEQUENCE_CONTEXT seq_ctx);
 extern void Game_End(void);