+++ src/tr1/game/game/game.c
@@ -60,6 +60,7 @@ void Game_ProcessInput(void)
 bool Game_Start(const GF_LEVEL *const level, const GF_SEQUENCE_CONTEXT seq_ctx)
 {
     Game_SetCurrentLevel(level);
+    Game_FadeToBlack(-1);
 
     g_OverlayFlag = 1;
     Camera_Initialise();