+++ src/tr2/game/game.c
@@ -27,6 +27,7 @@
 bool Game_Start(const GF_LEVEL *const level, const GF_SEQUENCE_CONTEXT seq_ctx)
 {
     Game_SetCurrentLevel(level);
+    Game_FadeToBlack(-1);
 
     g_OverlayStatus = 1;
     Camera_Initialise();
@@ -174,6 +175,9 @@ void Game_Draw(bool draw_overlay)
     } else {
         Overlay_HideGameInfo();
     }
+    Output_DrawPolyList();
+    Game_DrawFade();
+    Output_DrawPolyList();
 }
 
 void Game_ProcessInput(void)